XeTeX 0.9997.1.
git-svn-id: https://svn.code.sf.net/p/miktex/code/miktex/trunk@1848 3fc76331-e2db-4b81-af9c-3600f57961af

+++ Programs/TeXAndFriends/xetex/XeTeXFontMgr.cpp
@@ -44,8 +44,8 @@ const double M_PI = 3.14159265358979323846;
 
 #include "XeTeXswap.h"
 
-#include "Features.h"
-#include "GlyphPositioningTables.h"
+#include "layout/ICUFeatures.h"
+#include "layout/GlyphPositioningTables.h"
 
 #include "sfnt.h"
 
@@ -594,7 +594,7 @@ XeTeXFontMgr::getOpSizeRecAndStyleFlags(Font* theFont)
 
 		const POSTTable* postTable = (const POSTTable*)getFontTablePtr(font, LE_POST_TABLE_TAG);
 		if (postTable != NULL) {
-			theFont->slant = (int)(1000 * (tan(Fix2X(-SWAP(postTable->italicAngle)) * M_PI / 180.0)));
+			theFont->slant = (int)(1000 * (tan(Fix2X(-SWAP(UInt32(postTable->italicAngle))) * M_PI / 180.0)));
 		}
 		deleteFont(font);
 	}
